change servers.initAll to entryPoints.iniAll and add lb entrypoint in it
