+++ lib/api/core/cluster/clusterManager.js
@@ -2,8 +2,8 @@ function ClusterManager (kuzzle) {
 
   this.kuzzle = kuzzle;
 
+  // When the lb broker connection is open, we send to the LB this Kuzzle HTTP port configuration
   kuzzle.services.list.lbBroker.on('open', () => {
-
     kuzzle.services.list.lbBroker.send(JSON.stringify({
       data: {
         httpPort: this.context.httpPort
@@ -11,32 +11,7 @@ function ClusterManager (kuzzle) {
       webSocketId: null,
       room: 'httpPortInitialization'
     }));
-
-    kuzzle.services.list.lbBroker.join('request', this.onRequest.bind(this));
-    kuzzle.services.list.lbBroker.join('connection', this.onConnection.bind(this));
-    kuzzle.services.list.lbBroker.join('disconnect', this.onDisconnect.bind(this));
-    kuzzle.services.list.lbBroker.join('error', this.onDisconnect.bind(this));
   });
 }
-
-ClusterManager.prototype.onRequest = function (message) {
-  this.kuzzle.funnel.execute(message.data, message.connection, (error, responseObject) => {
-    this.kuzzle.services.list.lbBroker.send(JSON.stringify({
-      requestId: responseObject.requestId,
-      data: responseObject.toJson(),
-      connection: message.connection,
-      room: 'response'
-    }));
-    }
-  );
-};
-
-ClusterManager.prototype.onConnection = function (message) {
-  this.kuzzle.router.newConnection(message.connection.type, message.connection.id);
-};
-
-ClusterManager.prototype.onDisconnect = function (message) {
-  this.kuzzle.router.removeConnection(message.connection);
-};
 
 module.exports = ClusterManager;
