+++ lib/api/core/entryPoints/http.js
@@ -1,20 +1,12 @@
 var
   http = require('http');
 
-
-module.exports = {
-  initAll: function (kuzzle, params) {
-    runHttpServer(kuzzle, params);
-    runMQListener(kuzzle, params);
-  }
-};
-
 /**
  * Run a HTTP server and will use the router object to handle incoming requests
  * @param kuzzle
  * @param params
  */
-function runHttpServer (kuzzle, params) {
+module.exports = function runHttpServer (kuzzle, params) {
   var
     port = params.httpPort,
     server;
@@ -29,14 +21,4 @@ function runHttpServer (kuzzle, params) {
   kuzzle.pluginsManager.trigger('server:httpStarted', 'Starting: HTTP server on port ' + port);
 
   return server;
-}
-
-/**
- * Asks the router controller to start listening to messages coming from RabbitMQ
- * @param kuzzle
- */
-function runMQListener (kuzzle) {
-  kuzzle.router.routeMQListener();
-
-  kuzzle.pluginsManager.trigger('server:mqStarted', 'Starting: MQ listener');
-}
+};
