+++ lib/api/start.js
@@ -6,7 +6,7 @@
 var
   q = require('q'),
   packageInfo = require('../../package.json'),
-  servers = require('./core/servers'),
+  entryPoints = require('./core/entryPoints'),
   HotelClerk = require('./core/hotelClerk'),
   Notifier = require('./core/notifier'),
   Statistics = require('./core/statistics'),
@@ -109,16 +109,16 @@ module.exports = function start (params, feature) {
         // Worker response listener
         this.workerListener = new WorkerListener(this, this.config.queues.workerWriteResponseQueue);
 
+        // Create and init the Cluster
+        this.clusterManager = new ClusterManager(this);
+
         if (!feature.dummy) {
           // Initialize hooks
           this.hooks.init();
 
-          // Starts the servers in charge of listening to client queries (HTTP, MQ or WebSocket)
-          servers.initAll(this, params);
+          // Init all entry points in charge of listening to client queries (HTTP, Load balancer, MQ)
+          entryPoints.initAll(this, params);
         }
-
-        // Create and init the Cluster
-        this.clusterManager = new ClusterManager(this);
       }
 
       // Start a single set of workers
