+++ lib/config/index.js
@@ -23,6 +23,6 @@ module.exports = function (params) {
     internalIndex: params.internalIndex,
     request: params.request,
     cluster: params.cluster,
-    httpPort: process.env.KUZZLE_HTTP_PORT || params.httpPort || 7511
+    httpPort: params.httpPort
   };
 };
